+++ src/main/scala-2.11/markets/mutable/matching/CDAMatchingEngine.scala
@@ -13,8 +13,9 @@ WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 See the License for the specific language governing permissions and
 limitations under the License.
 */
-package markets.engines
+package markets.mutable.matching
 
+import markets.generic.Matching
 import markets.mutable.orderbooks.PriorityOrderBook
 import markets.orders.limit.LimitOrder
 import markets.orders.market.{MarketAskOrder, MarketBidOrder}
@@ -153,9 +154,9 @@ class CDAMatchingEngine(initialPrice: Long, tradable: Tradable)
     }
   }
 
-  protected[engines] val askOrderBook = PriorityOrderBook[AskOrder](tradable)(askOrdering)
+  protected[matching] val askOrderBook = PriorityOrderBook[AskOrder](tradable)(askOrdering)
 
-  protected[engines] val bidOrderBook = PriorityOrderBook[BidOrder](tradable)(bidOrdering)
+  protected[matching] val bidOrderBook = PriorityOrderBook[BidOrder](tradable)(bidOrdering)
 
   /* Cached value of most recent transaction price for internal use only. */
   private[this] var mostRecentPrice = initialPrice
