+++ src/test/scala-2.11/markets/mutable/matching/CDAMatchingEngineSpec.scala
@@ -13,9 +13,10 @@ WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 See the License for the specific language governing permissions and
 limitations under the License.
 */
-package markets.engines
+package markets.mutable.matching
 
 import markets.MarketsTestKit
+import markets.generic.Matching
 import markets.orders.limit.{LimitAskOrder, LimitBidOrder}
 import markets.orders.market.{MarketAskOrder, MarketBidOrder}
 import org.scalatest.{FeatureSpec, GivenWhenThen, Matchers}
